backend/requirements.txt
# E:\curamind\requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
pandas==2.2.3
scikit-learn>=1.5.2
xgboost==2.0.1
joblib==1.3.2
python-dotenv==1.0.0
openai==1.1.0
eli5
PyJWT>=2.10.1
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
email-validator>=2.0.0
pymongo>=4.6.0
psycopg2-binary>=2.9.9

vaderSentiment>=3.3.2